Lots of minor changes
