+++ web/quiz.html
@@ -3,6 +3,7 @@
         <meta charset="UTF-8">
         <meta name="viewport" content="width=device-width, initial-scale=1.0">
         <link rel="stylesheet" href="styles.css">
+        <script src="common.js"></script>
         <title>CleverDecks - Quiz</title>
     </head>
     <body>
